apps/api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pip==25.1.1
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-a2a>=0.5.9
pygments==2.19.1
pyjwt==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
